Refactor step executor and common step
